Move different views into separate files
